+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+- Fix joanie's course run link to LMS course in the syllabus page.
 - Fix enrollment cache not invalided after buying certificate product.
 - Fix a typo on ContractStatus component